feat(operator): author A&P confirmed settings + carry skill settings through parser and projection (#2005)
Christa's confirmed numbers (correspondence 09, 2026-07-23) authored on the
ashton-price seat: client-verification-tracker escalate_after_attempts: 3,
medical-chronology-maintainer treatment_gap_flag_days: 45. chase_cadence_days
stays deliberately unauthored — the letter commits a cadence "you set per
matter" (a working-session firm input); the skill fail-closes until then.

Authoring surfaced a wiring gap: the canonical TS parser DROPPED skills[]
.settings entirely (the overlay's translate consumes them, so the runtime
saw them but every ss-side consumer — validation, projection, portal — was
blind). Fixed end to end:

- PersonaSkill.settings: scalar-only map, mirroring the overlay's
  _skill_settings_block with one deliberate hardening: non-scalar values are
  a validation ERROR here (the overlay silently drops them, so accepting one
  would author a value the runtime never sees)
- customer_configs projection carries settings into personas_json (portal is
  the config source of truth to the client, correspondence 08)
- sections-personas.ts split at the 500-line ceiling: skill-entry +
  entitlement checks now live in sections-persona-skills.ts

Commitments gate (f) pins both values against correspondence 09 and asserts
chase_cadence_days stays unauthored. Parser + projection tests cover scalar
accept / nested reject / byte-stable absence / A&P pair round-trip.

Closes #2005.
